requirements.txt
flask==3.0.3
flask-socketio==5.4.1
flask-cors==5.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.8.2
requests==2.32.3
httpx==0.27.0
beautifulsoup4==4.12.3
playwright==1.47.0
pytest==9.0.2
pytest-asyncio==1.3.0
hypothesis==6.98.0
openai==2.15.0
jsonschema==4.23.0
imapclient==3.0.1
boto3==1.35.0
certifi
